backend/requirements.txt
aiofiles==0.8.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
attrs==21.4.0
bcrypt==3.2.0
boto3==1.21.44
botocore==1.24.44
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
databases==0.5.5
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
iniconfig==1.1.1
jmespath==1.0.0
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
python-decouple==3.6
python-multipart==0.0.5
requests==2.27.1
rfc3986==1.5.0
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.34
starlette==0.17.1
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
zipp==3.7.0
